start reworking xlibc

+++ xlibc/include/sys/types.h
@@ -7,7 +7,7 @@
 typedef uint64_t blkcnt_t;
 typedef uint64_t blksize_t;
 typedef uint64_t clock_t;
-typedef uint64_t clockid_t;
+typedef int clockid_t;
 typedef uint64_t dev_t;
 typedef uint64_t fsblkcnt_t;
 typedef uint64_t fsfilcnt_t;
@@ -17,7 +17,7 @@ typedef uint64_t ino_t;
 typedef uint64_t key_t;
 typedef uint64_t mode_t;
 typedef uint64_t nlink_t;
-typedef uint64_t off_t;
+typedef int64_t off_t;
 typedef uint64_t pid_t;
 typedef uint64_t pthread_attr_t;
 typedef uint64_t pthread_barrier_t;
